% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_page.R
\name{extract_nes_page}
\alias{extract_nes_page}
\title{extract_nes_page}
\usage{
extract_nes_page(nes_file, nes_page)
}
\arguments{
\item{nes_file}{file.path to NES pdf}

\item{nes_page}{numeric page number}
}
\description{
extract_nes_page
}
\examples{
\dontrun{
nes_file <- system.file("extdata/national-eutrophication-survey_477.PDF",
                     package = "nesR")
nes_page <- 11
tif_clean <- extract_nes_page(nes_file, nes_page)
}
}
