% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_page_list.R
\name{generate_page_list}
\alias{generate_page_list}
\title{generate_page_list}
\usage{
generate_page_list(pdfsource, outdir = ".")
}
\arguments{
\item{pdfsource}{numeric designation of the nes pdf source}

\item{outdir}{file.path to output directory}
}
\description{
generate_page_list
}
\examples{
\dontrun{
generate_page_list(477)
}
}

