% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residence-time_units.R
\name{get_residence_units}
\alias{get_residence_units}
\title{get_residence_units}
\usage{
get_residence_units(f_name)
}
\arguments{
\item{f_name}{character file name of raw data}
}
\description{
get_residence_units
}
\examples{
\dontrun{
f_name <- "474/100.csv"
get_residence_units(f_name)
}
}
