% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nes_get}
\alias{nes_get}
\title{nes_get}
\usage{
nes_get(nes_file, nes_page)
}
\arguments{
\item{nes_file}{file.path to NES pdf}

\item{nes_page}{numeric page number}
}
\description{
nes_get
}
\examples{
\dontrun{
nes_file <- system.file("extdata/national-eutrophication-survey_477.PDF",
                     package = "nesR")
nes_page <- 11
nes_get(nes_file, nes_page)
}
}

