% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\name{ocr_nes_page}
\alias{ocr_nes_page}
\title{ocr_nes_page}
\usage{
ocr_nes_page(tif_clean)
}
\arguments{
\item{tif_clean}{file.path to tif image for ocr}
}
\description{
ocr_nes_page
}
\examples{
\dontrun{
nes_file <- system.file("extdata/national-eutrophication-survey_477.PDF",
                     package = "nesR")
nes_page <- 11
tif_clean <- extract_nes_page(nes_file, nes_page)
ocr_nes_page(tif_clean)
}
}

