% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_nes}
\alias{parse_nes}
\title{parse_nes}
\usage{
parse_nes(tif_clean)
}
\arguments{
\item{tif_clean}{file.path to tif image for ocr}
}
\description{
parse_nes
}
\examples{
\dontrun{
parse_nes(tif_clean)
}
}

