% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_nes}
\alias{parse_nes}
\title{parse_nes}
\usage{
parse_nes(ocr_txt)
}
\arguments{
\item{ocr_txt}{character vector of ocr'd text}
}
\description{
parse_nes
}
\examples{
\dontrun{
parse_nes(read.csv(system.file(
"extdata/477_page-012.csv", package = "nesR"),
 stringsAsFactors = FALSE)[,1])
}
}
