% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_nes}
\alias{parse_nes}
\title{parse_nes}
\usage{
parse_nes(ocr_txt)
}
\arguments{
\item{ocr_txt}{character vector of ocr'd text}
}
\description{
parse_nes
}
\examples{
\dontrun{
nes_file <- system.file("extdata/national-eutrophication-survey_477.PDF",
                     package = "nesR")
nes_page <- 11
tif_clean <- extract_nes_page(nes_file, nes_page)
raw_txt <- ocr_nes_page(tif_clean)
parse_nes(raw_txt)
}
}

