% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_nes}
\alias{parse_nes}
\title{parse_nes}
\usage{
parse_nes(tif_clean)
}
\arguments{
\item{tif_clean}{file.path to tif image for ocr}
}
\description{
parse_nes
}
\examples{
\dontrun{
nes_file <- system.file("extdata/national-eutrophication-survey_1975.PDF",
                     package = "nesR")
nes_page <- 11
tif_clean <- extract_nes_page(nes_file, nes_page)
parse_nes(tif_clean)
}
}

